No routes are load in CamelContext
Issue: OHFJIRA-68
